WebSockets/tests/submissions/Opera/interfaces/WebSocket/readyState/007.html

<!doctype html>
<title>WebSockets: getting readyState in closing</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':80/ws/echo');
  ws.onopen = function(e) {
    
    ws.onclose = debug;
    ws.close();
    assert_equals(ws.readyState, ws.CLOSING);
    end();
  }
  ws.onerror = ws.onmessage = ws.onclose = t.step_func(function() {assert_unreached()});
});
</script>
